+++ src/behaviors/MuxAssetBehavior.php
@@ -8,10 +8,11 @@ use craft\web\View;
 
 use Twig\Markup;
 
-use vaersaagod\muxmate\helpers\MuxApiHelper;
+use vaersaagod\muxmate\MuxMate;
+use vaersaagod\muxmate\helpers\SignedUrlsHelper;
+use vaersaagod\muxmate\models\MuxPlaybackId;
 use vaersaagod\muxmate\helpers\MuxMateHelper;
 
-use vaersaagod\muxmate\MuxMate;
 use yii\base\Behavior;
 
 class MuxAssetBehavior extends Behavior
@@ -23,7 +24,15 @@ class MuxAssetBehavior extends Behavior
      */
     public function isMuxVideo(): bool
     {
-        return !empty($this->getMuxPlaybackId());
+        return !empty($this->getMuxAssetId());
+    }
+
+    /**
+     * @return string|null
+     */
+    public function getMuxStatus(): ?string
+    {
+        return MuxMateHelper::getMuxStatus($this->owner);
     }
 
     /**
@@ -35,43 +44,48 @@ class MuxAssetBehavior extends Behavior
     }
 
     /**
+     * @param string|null $policy
      * @return string|null
+     * @throws \yii\base\InvalidConfigException
      */
-    public function getMuxStreamUrl(): ?string
+    public function getMuxStreamUrl(?string $policy = null): ?string
     {
-        if (
-            !$this->owner instanceof Asset ||
-            !$playbackId = MuxMateHelper::getMuxPlaybackId($this->owner)
-        ) {
-            return null;
-        }
-        return MuxApiHelper::getStreamUrl($playbackId);
+        return MuxMateHelper::getMuxStreamUrl($this->owner, $policy);
     }
 
     /**
-     * @param array $params
+     * @param array|null $params
+     * @param string|null $policy
      * @return string|Markup
+     * @throws \yii\base\InvalidConfigException
      */
-    public function getMuxVideo(array $params = []): string|Markup
+    public function getMuxVideo(?array $params = null, ?string $policy = null): string|Markup
     {
-        if (
-            !$this->owner instanceof Asset ||
-            !MuxMateHelper::getMuxPlaybackId($this->owner)
-        ) {
+        if (!$this->owner instanceof Asset) {
             return '';
         }
+        $settings = MuxMate::getInstance()->getSettings();
+        $policy = $policy ?? $settings->defaultPolicy;
+        /** @var MuxPlaybackId|null $playbackId */
+        $playbackId = $this->getMuxPlaybackId($policy);
+        if (empty($playbackId)) {
+            return '';
+        }
+        $token = SignedUrlsHelper::getToken($playbackId, SignedUrlsHelper::SIGNED_URL_AUDIENCE_VIDEO, null, $this->getMuxVideoDuration());
         try {
-            $settings = MuxMate::getInstance()->getSettings();
             $muxVideoUrl = $settings->muxVideoUrl;
             $inline = $params['inline'] ?? null;
             $lazyload = $params['lazyload'] ?? $settings->lazyloadMuxVideo;
             $nonce = $settings->scriptSrcNonce;
             $html = Template::raw(\Craft::$app->getView()->renderTemplate('_muxmate/_mux-video.twig', [
                 'video' => $this->owner,
+                'playbackId' => $playbackId,
+                'token' => $token,
                 'muxVideoUrl' => $muxVideoUrl,
                 'inline' => $inline,
                 'lazyload' => $lazyload,
                 'nonce' => $nonce,
+                'policy' => $policy,
             ], View::TEMPLATE_MODE_CP));
         } catch (\Throwable $e) {
             \Craft::error($e, __METHOD__);
@@ -81,41 +95,49 @@ class MuxAssetBehavior extends Behavior
     }
 
     /**
-     * @param string $quality
+     * @param string|null $quality "high", "medium" or "low"
+     * @param string|null $policy "public" or "signed"
      * @param bool $download
      * @param string|null $filename
      * @return string|null
-     * @throws \Exception
+     * @throws \yii\base\InvalidConfigException
      */
-    public function getMuxMp4Url(string $quality = 'high', bool $download = false, ?string $filename = null): ?string
+    public function getMuxMp4Url(?string $quality = null, ?string $policy = null, bool $download = false, ?string $filename = null): ?string
     {
-        if (
-            !$this->owner instanceof Asset ||
-            !$playbackId = MuxMateHelper::getMuxPlaybackId($this->owner)
-        ) {
-            return null;
-        }
-        $qualities = ['high', 'medium', 'low'];
-        if (!in_array($quality, $qualities)) {
-            throw new \Exception("Invalid quality \"$quality\" (needs to be one of " . implode(', ', $qualities) . ')');
-        }
-        $muxData = MuxMateHelper::getMuxData($this->owner) ?? [];
-        $staticRenditions = $muxData['static_renditions'] ?? [];
-        if (($staticRenditions['status'] ?? null) !== 'ready') {
-            return null;
-        }
-        $availableQualities = [];
-        foreach ($muxData['static_renditions']['files'] ?? [] as $staticRendition) {
-            $availableQualities[] = explode('.', $staticRendition['name'])[0];
-        }
-        if (empty($availableQualities)) {
-            return null;
-        }
-        $availableQualities = array_values(array_intersect($qualities, $availableQualities));
-        if (!in_array($quality, $availableQualities)) {
-            $quality = $availableQualities[0];
-        }
-        return MuxApiHelper::getMp4Url($playbackId, $quality, $download, $filename);
+        return MuxMateHelper::getMuxMp4Url($this->owner, $quality, $policy, $download, $filename);
+    }
+
+    /**
+     * See https://docs.mux.com/guides/video/get-images-from-a-video for params
+     * @param array|null $params
+     * @param string|null $policy
+     * @return string|null
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function getMuxImageUrl(?array $params = null, ?string $policy = null): ?string
+    {
+        return MuxMateHelper::getMuxImageUrl($this->owner, $params, $policy);
+    }
+
+    /**
+     * See https://docs.mux.com/guides/video/get-images-from-a-video#get-an-animated-gif-from-a-video for params
+     *
+     * @param array|null $params
+     * @param string|null $policy
+     * @return string|null
+     * @throws \yii\base\InvalidConfigException
+     */
+    public function getMuxGifUrl(?array $params = null, ?string $policy = null): ?string
+    {
+        return MuxMateHelper::getMuxGifUrl($this->owner, $params, $policy);
+    }
+
+    /**
+     * @return float|null
+     */
+    public function getMuxVideoDuration(): ?float
+    {
+        return MuxMateHelper::getMuxVideoDuration($this->owner);
     }
 
     /**
@@ -123,59 +145,7 @@ class MuxAssetBehavior extends Behavior
      */
     public function getStaticRenditions(): ?array
     {
-        if (!$this->owner instanceof Asset) {
-            return null;
-        }
-        $muxData = MuxMateHelper::getMuxData($this->owner) ?? [];
-        $staticRenditions = $muxData['static_renditions'] ?? [];
-        if (($staticRenditions['status'] ?? null) !== 'ready') {
-            return null;
-        }
-        $staticRenditionsByQuality = [];
-        foreach ($staticRenditions['files'] ?? [] as $staticRendition) {
-            $quality = explode('.', $staticRendition['name'])[0];
-            $staticRenditionsByQuality[$quality] = $staticRendition;
-        }
-        if (empty($staticRenditionsByQuality)) {
-            return null;
-        }
-        return $staticRenditionsByQuality;
-    }
-
-    /**
-     * See https://docs.mux.com/guides/video/get-images-from-a-video for params
-     *
-     * @param array $params
-     * @return string|null
-     */
-    public function getMuxImageUrl(array $params = []): ?string
-    {
-        if (
-            !$this->owner instanceof Asset ||
-            !MuxMateHelper::getMuxPlaybackId($this->owner) ||
-            MuxMateHelper::getMuxStatus($this->owner) !== 'ready'
-        ) {
-            return null;
-        }
-        return MuxApiHelper::getImageUrl(MuxMateHelper::getMuxPlaybackId($this->owner), $params);
-    }
-
-    /**
-     * See https://docs.mux.com/guides/video/get-images-from-a-video#get-an-animated-gif-from-a-video for params
-     *
-     * @param array $params
-     * @return string|null
-     */
-    public function getMuxGifUrl(array $params = []): ?string
-    {
-        if (
-            !$this->owner instanceof Asset ||
-            !MuxMateHelper::getMuxPlaybackId($this->owner) ||
-            MuxMateHelper::getMuxStatus($this->owner) !== 'ready'
-        ) {
-            return null;
-        }
-        return MuxApiHelper::getGifUrl(MuxMateHelper::getMuxPlaybackId($this->owner), $params);
+        return MuxMateHelper::getStaticRenditions($this->owner);
     }
 
     /**
@@ -190,14 +160,16 @@ class MuxAssetBehavior extends Behavior
     }
 
     /**
+     * @param string|null $policy
      * @return string|null
+     * @throws \yii\base\InvalidConfigException
      */
-    public function getMuxPlaybackId(): ?string
+    public function getMuxPlaybackId(?string $policy = null): ?string
     {
         if (!$this->owner instanceof Asset) {
             return null;
         }
-        return MuxMateHelper::getMuxPlaybackId($this->owner);
+        return MuxMateHelper::getMuxPlaybackId($this->owner, $policy);
     }
 
     /**
@@ -209,17 +181,6 @@ class MuxAssetBehavior extends Behavior
             return null;
         }
         return MuxMateHelper::getMuxData($this->owner);
-    }
-
-    /**
-     * @return string|null
-     */
-    public function getMuxStatus(): ?string
-    {
-        if (!$this->owner instanceof Asset) {
-            return null;
-        }
-        return MuxMateHelper::getMuxStatus($this->owner);
     }
 
     /**