+++ src/config.php
@@ -3,7 +3,10 @@
 return [
     'muxAccessTokenId' => null,
     'muxSecretKey' => null,
+    'muxSigningKey' => null,
     'muxVideoUrl' => null,
+    'defaultPolicy' => null,
+    'defaultMp4Quality' => null,
     'lazyloadMuxVideo' => false,
     'scriptSrcNonce' => null,
     'volumes' => null,