+++ src/models/Settings.php
@@ -17,9 +17,15 @@ class Settings extends Model
 
     public ?string $muxSecretKey = null;
 
+    public ?MuxSigningKey $muxSigningKey = null;
+
     public string|bool|null $muxPlayerUrl = null;
 
     public string|bool|null $muxVideoUrl = null;
+
+    public string $defaultPolicy = 'public';
+
+    public string $defaultMp4Quality = 'high';
 
     public bool $lazyloadMuxVideo = false;
 
@@ -36,6 +42,18 @@ class Settings extends Model
         $values['muxVideoUrl'] = $values['muxVideoUrl'] ?? null;
         if ($values['muxVideoUrl'] !== false) {
             $values['muxVideoUrl'] = App::parseEnv($values['muxVideoUrl']) ?: MuxMate::MUX_VIDEO_URL;
+        }
+        if (!empty($values['muxSigningKey'])) {
+            $values['muxSigningKey'] = new MuxSigningKey([
+                'id' => $values['muxSigningKey']['id'] ?? null,
+                'privateKey' => $values['muxSigningKey']['privateKey'] ?? null,
+            ]);
+        }
+        if (array_key_exists('defaultPolicy', $values) && empty($values['defaultPolicy'])) {
+            unset($values['defaultPolicy']);
+        }
+        if (array_key_exists('defaultMp4Quality', $values) && empty($values['defaultMp4Quality'])) {
+            unset($values['defaultMp4Quality']);
         }
         parent::setAttributes($values, $safeOnly);
     }